double scaling for gantt charts

+++ oscar-cp/src/main/scala/oscar/cp/scheduling/visual/VisualGanttChart.scala
@@ -41,7 +41,7 @@ class VisualGanttChart(starts: Array[CPIntVar], durations: Array[CPIntVar], ends
   private val makespanLine: VisualLine = VisualLine(this, 0, 0, 0, 0)
   makespanLine.outerCol = Color.RED;
 
-  def update(xScale: Int, yScale: Int) {
+  def update(xScale: Double, yScale: Double) {
 
     for (i <- 0 until starts.size) {
